Added module for visualization functions
